Update raspiCamera.py
added anti-shake to the code making the raspcamera.py obsolete. Their is no need for the camera to take a continuous photo instead of having a variable to control when the code to then take a picture. If the U2 team really wants the time and date info with the picture we could change the naming of it or make it similar by adding and index.txt and printing the variable and the time that the picture was take.